migrations/requirements.txt
annotated-types==0.6.0
psycopg2-binary
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
typing_extensions==4.8.0
